Support selective authentication based on the request
A user can supply an auth_required_callback which when passed a copy of the wsgi
environment, will return True of False to indicate whether or not authentication
is required for the given request.

This allows the user to require selective authentication based on anything in
the wsgi environment, which includes things like the request method, url, and
even the data. Special care MUST be taken to not consume the data in a
destructive manner. If all of the data is consumed, then the authenticated
application may hang when it attempts to read the data.