Consume all data on a request before sending 401/403
This avoids problems that some clients have when they get an unexpected
and premature close from the server.

RFC2616: If an origin server receives a request that does not include an
Expect request-header field with the "100-continue" expectation, the
request includes a request body, and the server responds with a final
status code before reading the entire request body from the transport
connection, then the server SHOULD NOT CLOSE the transport connection until
it has read the entire request, or until the client closes the connection.
Otherwise, the client might not reliably receive the response message.
However, this requirement is not be construed as preventing a server from
defending itself against denial-of-service attacks, or from badly broken
client implementations.