Don't needlessly limit to service resolved at init time.
Previously, when the optional `hostname` argument was not passed to
__init__, it would fall back to `socket.gethostname()`.

__init__ would then populate `self.service` based on the resulting
`hostname` value, and this same `self.service` value would then always
be passed to `authGSSServerInit`.

However, pre-specifying a service when calling `authGSSServerInit`
needlessly constrains what service names can work and breaks any clients
that attempt to access the service by another name, even if a key for
that service name exists in the keytab by the time authentication
happens.

Now, when `hostname` is not passed to to __init__, empty string is passed
to `authGSSServerInit` to allow clients accessing the service by another
name (for which a key exists in the keytab by the time it is called) to
do so successfully.

Fixes #3.